# Enhance Performance by Adding `index_t` Template Parameter