Select `int64_t` or `int` data types at the start of GPU kernel computation, using templates to pass the chosen type. This optimization improves performance for smaller shapes.


#### Performance Improvement Reasons:
1. **Better Hardware Support**: GPUs handle 32-bit operations more efficiently than 64-bit.
2. **Reduced Memory Usage**: 32-bit integers use less memory, enhancing bandwidth and cache efficiency.
3. **Faster Instructions**: 32-bit operations require fewer instructions, speeding up computation.