@EikanWang 2025-04-26
@chunhuanMeng , pls. share the detailed performance improvement number here. Meanwhile, the pr description is meaningless. Pls elaborate on `index_t` and why `index_t` can improve performance.