@chunhuanMeng 2025-04-28
Performance Comparison
dtype               | shape                                                                 | original/us | optimized/us |
|---------------------|----------------------------------------------------------------------|-------------|--------------|
| fp16                | ['float16[16, 24, 112, 112]', [3, 3], [2, 2], [0, 0], False, True, None] | 45.904      | 38.368       |
| fp16                | ['float16[16, 1984, 7, 7]', [3, 2], [2, 1], [0, 0], False, True, None] | 30.336      | 24.664       |
| fp16                | ['float16[64, 1024, 112, 112]', [6, 6], [4, 4], [0, 0], False, True, None] | 2970.816    | 2686.872     |
| fp16                | ['float16[16, 2048, 224, 224]', [3, 3], [2, 2], [0, 0], False, True, None] | 12278.568   | 10279.36     |
| fp16_ChannelsLast   | ['float16[16, 24, 112, 112]', [3, 3], [2, 2], [0, 0], False, True, None] | 62.472      | 42.744       |
| fp16_ChannelsLast   | ['float16[16, 1984, 7, 7]', [3, 2], [2, 1], [0, 0], False, True, None] | 31.128      | 25.272       |
| fp16_ChannelsLast   | ['float16[64, 1024, 112, 112]', [6, 6], [4, 4], [0, 0], False, True, None] | 8104.144    | 6541.616     |
| fp16_ChannelsLast   | ['float16[16, 2048, 224, 224]', [3, 3], [2, 2], [0, 0], False, True, None] | 23269.616   | 17246.784    |
| fp32                | ['float32[16, 24, 112, 112]', [3, 3], [2, 2], [0, 0], False, True, None] | 49.616      | 38.816       |
| fp32                | ['float32[16, 1984, 7, 7]', [3, 2], [2, 1], [0, 0], False, True, None] | 28.032      | 20.392       |
| fp32                | ['float32[64, 1024, 112, 112]', [6, 6], [4, 4], [0, 0], False, True, None] | 3353.68     | 3324.376     |
| fp32                | ['float32[16, 2048, 224, 224]', [3, 3], [2, 2], [0, 0], False, True, None] | 12431.688   | 10453.808    |
| fp32_ChannelsLast   | ['float32[16, 24, 112, 112]', [3, 3], [2, 2], [0, 0], False, True, None] | 69.2        | 50.416       |
| fp32_ChannelsLast   | ['float32[16, 1984, 7, 7]', [3, 2], [2, 1], [0, 0], False, True, None] | 36.056      | 28.848       |
| fp32_ChannelsLast   | ['float32[64, 1024, 112, 112]', [6, 6], [4, 4], [0, 0], False, True, None] | 8016.92     | 6475.76      |
| fp32_ChannelsLast   | ['float32[16, 2048, 224, 224]', [3, 3], [2, 2], [0, 0], False, True, None] | 22995.984   | 17047.776    |